refactor: extract media query hook
